Refactor instant hustle helpers into modular files
